Tintin platform cleanup. Display DMA
