+++ hw/platform/tintin/tintin.c
@@ -33,7 +33,6 @@
 static inline void _init_USART3();
 static int _debug_initialized;
 
-uint8_t hw_bluetooth_power_cycle(void);
 
 static const stm32_usart_config_t _u3_config = {
     .usart                = USART3,
@@ -55,43 +54,6 @@ static stm32_usart_t _usart3 = {
     230400
 };
 
-
-static const stm32_spi_config_t _spi2_config = {
-    .spi                  = SPI2,
-    .spi_periph_bus       = STM32_POWER_APB1,
-    .gpio_pin_miso_num    = 0,
-    .gpio_pin_mosi_num    = 15,
-    .gpio_pin_sck_num     = 13,
-    .gpio_ptr             = GPIOB,
-    .gpio_clock           = RCC_AHB1Periph_GPIOB,
-    .spi_clock            = RCC_APB1Periph_SPI2,
-    .af                   = GPIO_AF_SPI2,
-    .txrx_dir             = STM32_SPI_DIR_TX,
-    .crc_poly             = 7 /* Um */
-};
-
-static const stm32_dma_t _spi2_dma = {
-    .dma_clock            = RCC_AHB1Periph_DMA1,
-    .dma_tx_stream        = DMA1_Stream4,
-    .dma_rx_stream        = DMA1_Stream3,
-    .dma_tx_channel       = DMA_Channel_0,
-    .dma_rx_channel       = DMA_Channel_0,
-    .dma_irq_tx_pri       = 10,
-    .dma_irq_rx_pri       = 9,
-    .dma_irq_tx_channel   = DMA1_Stream4_IRQn,
-    .dma_irq_rx_channel   = DMA1_Stream3_IRQn,
-    .dma_tx_channel_flags = STM32_DMA_MK_FLAGS(4),
-    .dma_rx_channel_flags = STM32_DMA_MK_FLAGS(3),
-    .dma_tx_irq_flag      = DMA_IT_TCIF4,
-    .dma_rx_irq_flag      = DMA_IT_TCIF3
-};
-
-static stm32_spi_t _spi2 = {
-    &_spi2_config,
-    &_spi2_dma, /* dma */
-};
-
-//STM32_SPI_MK_TX_IRQ_HANDLER(&_spi2, 2, 5, _spi_tx_done)
 
 void debug_init() {
     _init_USART3();
@@ -212,90 +174,6 @@ stm32_backlight_config_t platform_backlight = {
   .rcc_tim = RCC_APB1Periph_TIM3
 };
 
-/* display */
-
-static uint8_t _display_fb[168][20];
-
-
-void hw_display_init() {
-    printf("tintin: hw_display_init\n");
-
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOB);
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOC);
-
-    GPIO_WriteBit(GPIOB, 1 << 12, 0);
-    GPIO_PinAFConfig(GPIOB, 1, GPIO_AF_TIM3);
-   
-    GPIO_InitTypeDef gpioinit;
-    
-    gpioinit.GPIO_Pin = (1 << 12);
-    gpioinit.GPIO_Mode = GPIO_Mode_OUT;
-    gpioinit.GPIO_Speed = GPIO_Speed_50MHz;
-    gpioinit.GPIO_OType = GPIO_OType_PP;
-    gpioinit.GPIO_PuPd = GPIO_PuPd_NOPULL;
-    GPIO_Init(GPIOB, &gpioinit);
-    
-    gpioinit.GPIO_Pin = (1 << 1);
-    gpioinit.GPIO_Mode = GPIO_Mode_AF;
-    gpioinit.GPIO_Speed = GPIO_Speed_50MHz;
-    gpioinit.GPIO_OType = GPIO_OType_OD;
-    gpioinit.GPIO_PuPd = GPIO_PuPd_NOPULL;
-    GPIO_Init(GPIOB, &gpioinit);
-
-    /* Set up the SPI controller, SPI2. */
-    stm32_spi_init_device(&_spi2);
-    
-    stm32_power_release(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOB);
-    stm32_power_release(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOC);
-}
-
-void hw_display_reset() {
-    printf("tintin: hw_display_reset\n");
-}
-
-void hw_display_start() {
-    printf("tintin: hw_display_start\n");
-}
-
-void hw_display_start_frame(uint8_t x, uint8_t y) {
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOB);
-    stm32_power_request(STM32_POWER_APB1, RCC_APB1Periph_SPI2);
-
-    printf("tintin: here we go, slowly blitting %d %d\n", x, y);
-    GPIO_WriteBit(GPIOB, 1 << 12, 1);
-    delay_us(7);
-    stm32_spi_write(&_spi2, 0x80);
-    for (int i = 0; i < 168; i++) {
-        stm32_spi_write(&_spi2, __RBIT(__REV(168-i)));
-        for (int j = 0; j < 18; j++)
-            stm32_spi_write(&_spi2, _display_fb[i][17-j]);
-        stm32_spi_write(&_spi2, 0);
-    }
-    stm32_spi_write(&_spi2, 0);
-    delay_us(7);
-    GPIO_WriteBit(GPIOB, 1 << 12, 0);
-
-    stm32_power_release(STM32_POWER_APB1, RCC_APB1Periph_SPI2);
-    stm32_power_release(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOB);
-
-    display_done_ISR(0);
-}
-
-uint8_t *hw_display_get_buffer(void) {
-    return (uint8_t *)_display_fb;
-}
-
-uint8_t hw_display_get_state() {
-    return 1;
-}
-
-
-// void scanline_rgb888pixel_to_frambuffer(UG_S16 x, UG_S16 y, UG_COLOR c) {
-//     if (c)
-//         _display_fb[y][17 - (x / 8)] |= (0x80 >> (x % 8));
-//     else
-//         _display_fb[y][17 - (x / 8)] &= ~(0x80 >> (x % 8));
-// }
 
 /* vibrate */
 
@@ -306,147 +184,6 @@ void hw_vibrate_enable(uint8_t enabled) {
 }
 
 
-#define JEDEC_READ 0x03
-#define JEDEC_RDSR 0x05
-#define JEDEC_IDCODE 0x9F
-#define JEDEC_DUMMY 0xA9
-#define JEDEC_WAKE 0xAB
-#define JEDEC_SLEEP 0xB9
-
-#define JEDEC_RDSR_BUSY 0x01
-
-#define JEDEC_IDCODE_MICRON_N25Q032A11 0x20BB16 /* bianca / qemu / ev2_5 */
-#define JEDEC_IDCODE_MICRON_N25Q064A11 0x20BB17 /* v1_5 */
-static uint8_t _hw_flash_txrx(uint8_t c) {
-    while (!(SPI1->SR & SPI_SR_TXE))
-        ;
-    SPI1->DR = c;
-    while (!(SPI1->SR & SPI_SR_RXNE))
-        ;
-    return SPI1->DR;
-}
-
-static void _hw_flash_enable(int i) {
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOA);
-
-    GPIO_WriteBit(GPIOA, 1 << 4, !i);
-    delay_us(1);
-    
-    stm32_power_release(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOA);
-}
-
-static void _hw_flash_wfidle() {
-    _hw_flash_enable(1);
-    _hw_flash_txrx(JEDEC_RDSR);
-    while (_hw_flash_txrx(JEDEC_DUMMY) & JEDEC_RDSR_BUSY)
-        ;
-    _hw_flash_enable(0);
-}
-
-
-void hw_flash_init(void) {
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOA);
-    
-    /* Set up the pins. */
-    GPIO_WriteBit(GPIOA, 1 << 4, 0); /* nCS */
-    GPIO_PinAFConfig(GPIOA, 5, GPIO_AF_SPI1);
-    GPIO_PinAFConfig(GPIOA, 6, GPIO_AF_SPI1);
-    GPIO_PinAFConfig(GPIOA, 7, GPIO_AF_SPI1);
-    
-    GPIO_InitTypeDef gpioinit;
-    
-    gpioinit.GPIO_Pin = (1 << 7) | (1 << 6);
-    gpioinit.GPIO_Mode = GPIO_Mode_AF;
-    gpioinit.GPIO_Speed = GPIO_Speed_50MHz;
-    gpioinit.GPIO_OType = GPIO_OType_PP;
-    gpioinit.GPIO_PuPd = GPIO_PuPd_NOPULL;
-    GPIO_Init(GPIOA, &gpioinit);
-
-    gpioinit.GPIO_Pin = (1 << 5);
-    gpioinit.GPIO_Mode = GPIO_Mode_AF;
-    gpioinit.GPIO_Speed = GPIO_Speed_50MHz;
-    gpioinit.GPIO_OType = GPIO_OType_PP;
-    gpioinit.GPIO_PuPd = GPIO_PuPd_DOWN;
-    GPIO_Init(GPIOA, &gpioinit);
-
-    gpioinit.GPIO_Pin = (1 << 4);
-    gpioinit.GPIO_Mode = GPIO_Mode_OUT;
-    gpioinit.GPIO_Speed = GPIO_Speed_50MHz;
-    gpioinit.GPIO_OType = GPIO_OType_PP;
-    gpioinit.GPIO_PuPd = GPIO_PuPd_UP;
-    GPIO_Init(GPIOA, &gpioinit);
-    
-    stm32_power_release(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOA);
-
-    /* Set up the SPI controller, SPI1. */
-    SPI_InitTypeDef spiinit;
-    
-    stm32_power_request(STM32_POWER_APB2, RCC_APB2Periph_SPI1);
-
-    SPI_I2S_DeInit(SPI1);
-    
-    spiinit.SPI_Direction = SPI_Direction_2Lines_FullDuplex;
-    spiinit.SPI_Mode = SPI_Mode_Master;
-    spiinit.SPI_DataSize = SPI_DataSize_8b;
-    spiinit.SPI_CPOL = SPI_CPOL_Low;
-    spiinit.SPI_CPHA = SPI_CPHA_1Edge;
-    spiinit.SPI_NSS = SPI_NSS_Soft;
-    spiinit.SPI_BaudRatePrescaler = SPI_BaudRatePrescaler_2;
-    spiinit.SPI_FirstBit = SPI_FirstBit_MSB;
-    spiinit.SPI_CRCPolynomial = 7 /* Um. */;
-    SPI_Init(SPI1, &spiinit);
-    SPI_Cmd(SPI1, ENABLE);
-    
-    /* In theory, SPI is up.  Now let's see if we can talk to the part. */
-    _hw_flash_enable(1);
-    _hw_flash_txrx(JEDEC_WAKE);
-    _hw_flash_enable(0);
-    delay_us(100);
-    
-    _hw_flash_wfidle();
-     
-    uint32_t part_id = 0;
-    
-    _hw_flash_enable(1);
-    _hw_flash_txrx(JEDEC_IDCODE);
-    part_id |= _hw_flash_txrx(JEDEC_DUMMY) << 16;
-    part_id |= _hw_flash_txrx(JEDEC_DUMMY) << 8;
-    part_id |= _hw_flash_txrx(JEDEC_DUMMY) << 0;
-    _hw_flash_enable(0);
-    
-    printf("tintin flash: JEDEC ID %08lx\n", part_id);
-
-    if (part_id != JEDEC_IDCODE_MICRON_N25Q032A11 && part_id != JEDEC_IDCODE_MICRON_N25Q064A11) {
-        panic("tintin flash: unsupported part ID");
-    }
-    
-    stm32_power_release(STM32_POWER_APB2, RCC_APB2Periph_SPI1);
-}
-
-
-void hw_flash_read_bytes(uint32_t addr, uint8_t *buf, size_t len) {
-    assert(addr < 0x1000000 && "address too large for JEDEC_READ command");
-    
-    stm32_power_request(STM32_POWER_APB2, RCC_APB2Periph_SPI1);
-
-    _hw_flash_wfidle();
-    
-    _hw_flash_enable(1);
-    _hw_flash_txrx(JEDEC_READ);
-    _hw_flash_txrx((addr >> 16) & 0xFF);
-    _hw_flash_txrx((addr >>  8) & 0xFF);
-    _hw_flash_txrx((addr >>  0) & 0xFF);
-    
-    /* XXX: could use DMA, conceivably */
-    for (int i = 0; i < len; i++)
-        buf[i] = _hw_flash_txrx(JEDEC_DUMMY);
-    
-    /* make sure we are fully clocked out before we drop enable */
-//     delay_us(100);
-    _hw_flash_enable(0);
-
-    stm32_power_release(STM32_POWER_APB2, RCC_APB2Periph_SPI1);
-}
 
 void ss_debug_write(const unsigned char *p, size_t len)
 {
@@ -459,78 +196,6 @@ void log_clock_disable() {
 }
 
 
-/* Bluetooth */
-static const stm32_usart_config_t _usart1_config = {
-    .usart                = USART1,
-    .flow_control_enabled = FLOW_CONTROL_ENABLED,
-    .usart_periph_bus     = STM32_POWER_APB2,
-    .gpio_pin_tx_num      = 9,
-    .gpio_pin_rx_num      = 10,
-    .gpio_pin_rts_num     = 12,
-    .gpio_pin_cts_num     = 11,
-    .gpio_ptr             = GPIOA,
-    .gpio_clock           = RCC_AHB1Periph_GPIOA,
-    .usart_clock          = RCC_APB2Periph_USART1,
-    .af                   = GPIO_AF_USART1,
-};
-
-/* dma tx: dma stream 7 chan 4, rx: stream 2 chan 4. 
- * irq tx: 6, rx: 8 */
-static const stm32_dma_t _usart1_dma = STM32_DMA_MK_INIT(RCC_AHB1Periph_DMA2, 2, 7, 2, 4, 4, 6, 8); 
-
-static stm32_usart_t _usart1 = {
-    &_usart1_config,
-    &_usart1_dma, /* dma */
-    115200 /* initial slow handshake */
-};
-
-
-static const stm32_bluetooth_config_t stm32_cc256x_config = {
-    .usart                = &_usart1,
-    .shutdown_power       = STM32_POWER_AHB1,
-    .shutdown_power_port  = RCC_AHB1Periph_GPIOA,
-    .shutdown_pin_num     = 3,
-    .shutdown_port        = GPIOA,
-};    
-
-
-/* IRQ handlers
- * TX: DMA 2 stream 7.
- * RX: DMA 2 stream 2. */
-STM32_CC256X_MK_IRQ_HANDLERS(&_usart1, 2, 7, 2)
-
-uint8_t hw_bluetooth_init(void)
-{
-    stm32_power_request(STM32_POWER_AHB1, RCC_AHB1Periph_GPIOA);
-    stm32_power_request(STM32_POWER_APB2, RCC_APB2Periph_USART1);
-    stm32_cc256x_init(&stm32_cc256x_config);
-    return 0;
-}
-
-inline void hw_bluetooth_clock_on(void)
-{
-    stm32_cc256x_clock_on();
-}
-
-inline uint8_t hw_bluetooth_power_cycle(void)
-{
-    return stm32_cc256x_power_cycle();
-}
-
-void hw_bluetooth_enable_cts_irq()
-{
-    stm32_cc256x_enable_cts_irq();
-}
-
-void hw_bluetooth_disable_cts_irq(void)
-{
-    stm32_cc256x_disable_cts_irq();
-}
-
-stm32_usart_t *hw_bluetooth_get_usart(void)
-{
-    return &_usart1;
-}
 
 
 void HardFault_Handler(uint32_t *sp)
